backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
anthropic>=0.28.0
Pillow>=10.3.0
piexif>=1.1.3
python-dotenv>=1.0.1
